Add negative space to blueprints on create
This is about half of what I want. I would like the distance of
exclusion to be configurable, but that would be pretty difficult. I
would also like to exclude inserters from interacting with inventories,
but that requires knowing the bounding box of rotateable entities, like
railgun turrets, which is very hard.