Fix flipping blueprints breaking directionality
This was a tough one. When a blueprint is flipped with V or H, any
entity is rotated 180 degrees as needed. But if it doesn't support
flipping, no change is needed.

But our entity stores directional data, and so needs to know if it was
flipped.

Unfortunately, the game does not let any entity be mirrored except for
crafting machines. And so we make our auto guy a crafting machine.

This makes no difference, as it will never be in the world, except in
ghost form for a tick.