Half-way fix small undo issue
The undo/redo support for negative-space is terrible. But the game does
not let us edit undo/redo history, except for adding an action to the
undo history, and removing an action from the undo/redo stack. This
means we have no way to add a "build entity" action to the redo stack.